+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## [Unreleased]
 
--   External dependencies can be added as optionals in configuration 
+-   External dependencies can be added as optionals in configuration
+-   Support for labels, tolerations and annotations in k8s
+-   Logging added to operators
 
 ## [0.4.0] - 2021-04-20